Change File structure of networking
