+++ TCAT/Core/Network/Base/NetworkManager.swift
